Added advanced file sorting and path handling
Layers are now classified as "Vector", "Raster" and "Other". Files are sorted into the appropriate subdirectories.

Relative paths are now turned into absolute paths and web URLs are skipped.

+++ R/qgis_filesort.R
@@ -2,11 +2,14 @@
 #'
 #' This function moves a QGIS project and all the source files for its layers
 #' into a single directory. Source paths, styling, definition queries,
-#' and metadata are kept intact.
+#' and metadata are kept intact. Layer files are sorted into appropriate
+#' subdirectories ("Vector", "Raster", "Other") based on file extention.
 #'
 #' @details
 #' This function processes a QGIS project file, copying all the associated
 #' layer files and adjusting their paths to be relative to the new directory.
+#' Layers are sorted based on their type.
+#'
 #' Metadata, styling, and definition queries are preserved during the operation.
 #'
 #' @param input_path Path to the .qgz QGIS project file.
@@ -22,6 +25,7 @@
 #' }
 #' @importFrom xml2 read_xml xml_find_all xml_text xml_set_text write_xml
 #' @importFrom utils unzip zip
+#' @importFrom tools file_path_sans_ext
 #' @export
 
 
@@ -30,6 +34,7 @@ qgis.filesort <- function(input_path, output_dir ="", remove_old = FALSE) {
   if (!file.exists(input_path) || !grepl("\\.qgz$", input_path)) {
     stop("Invalid path: The path does not exist or is not a .qgz file.")
   }
+
   # Get path components
   project_file <- basename(input_path)
   project_dir <- dirname(input_path)
@@ -39,9 +44,9 @@ qgis.filesort <- function(input_path, output_dir ="", remove_old = FALSE) {
   if (output_dir == ""){
     output_dir <- tools::file_path_sans_ext(input_path)
     # Create new folder
-    if (!file.exists(output_dir)) {
+    if (!dir.exists(output_dir)) {
       dir.create(output_dir)
-      print(paste("Folder created:", project_sans))
+      print(paste("Folder created:", output_dir))
     }
   }
 
@@ -60,47 +65,86 @@ qgis.filesort <- function(input_path, output_dir ="", remove_old = FALSE) {
   for (layer in project_layernodes){
     # Get path from xml
     layerpath_old <- xml_text(layer)
+    # Catch definition queries
     layerpath_query <- ""
     if (grepl("\\|", layerpath_old)){
-      print(paste("Definition Query caught in layer:", layerpath_old))
       layerpath_query <- paste0( "\\|", sub(".*\\|", "", layerpath_old))
       layerpath_old <- sub("\\|.*", "", layerpath_old)
     }
-    if(file.exists(layerpath_old)){
-      # Catch definition queries
+
+    # Catch some different path types
+    if (!grepl("^([a-zA-Z]:|/)", layerpath_old)) {
+      # Turn relative layer path into absolute path
+      layerpath_old <- file.path(project_dir, layerpath_old)
+      layerpath_old <- normalizePath(layerpath_old, winslash = "/", mustWork = FALSE)
+    }
+    else if (grepl("^https?://", layerpath_old, ignore.case = TRUE)) {
+      # Skip web URLs
+      next
+    }
+
+    # Check if file exists
+    if(file.exists(layerpath_old) && !dir.exists(layerpath_old)){
+      # Classify layer
+      if (grepl("\\.(tif|tiff|jpg|png|jpeg|bmp|gif|aux\\.xml|ovr|rrd|tfw|jgw|pgw|bpw|gfw|prj|vrt)$", layerpath_old, ignore.case = TRUE)) {
+        layerclass <- "Raster"
+      } else if (grepl("\\.(shp|geojson|json|gpkg|kml|gml|shx|dbf|prj|cpg|qpj|sbn|sbx|aih|xml|kmz|xsd)$", layerpath_old, ignore.case = TRUE)) {
+        layerclass <- "Vector"
+      } else {
+        layerclass <- "Other"
+      }
+      # Create class-based subdirectory if it doesn't exist yet
+      class_dir <- paste0(output_dir, "/", layerclass)
+      if (!dir.exists(class_dir)) {
+        dir.create(class_dir, recursive = TRUE)
+      }
+
       # Adjust path in xml
       layerpath_file <- basename(layerpath_old)
-      xml_set_text(layer, paste0("./", layerpath_file, layerpath_query))
+      xml_set_text(layer, paste0("./", layerclass, "/", layerpath_file, layerpath_query))
 
       # Get dir and filename without extension for the current layer
       layerpath_dir <- dirname(layerpath_old)
       layerpath_sans <- tools::file_path_sans_ext(layerpath_file)
-      print(layerpath_sans)
 
       # Find file and all corresponding metadata files, copy them to new location
       layerlist <- list.files(path = layerpath_dir, pattern = paste0("^", layerpath_sans, "\\..+$"), full.names = TRUE, recursive = FALSE)
       for (layerfile in layerlist){
-        layerpath_new <- paste0(output_dir, "/", basename(layerfile))
-        file.copy(layerfile, layerpath_new)
-        # Remove of file if requested and if file was successfully copied
-        if(remove_old){
-          if(file.exists(layerpath_new)){
+        if (layerclass == "Raster" && grepl("\\.(tif|tiff|jpg|png|jpeg|bmp|gif|aux\\.xml|ovr|rrd|tfw|jgw|pgw|bpw|gfw|prj|vrt)$", layerfile, ignore.case = TRUE)){
+          layerpath_new <- paste0(class_dir, "/", basename(layerfile))
+          file.copy(layerfile, layerpath_new)
+        }
+        else if (layerclass == "Vector" && grepl("\\.(shp|geojson|json|gpkg|kml|gml|shx|dbf|prj|cpg|qpj|sbn|sbx|aih|xml|kmz|xsd)$", layerfile, ignore.case = TRUE)){
+          layerpath_new <- paste0(class_dir, "/", basename(layerfile))
+          file.copy(layerfile, layerpath_new)
+        }
+        else if (layerclass == "Other"){
+          layerpath_new <- paste0(class_dir, "/", basename(layerfile))
+          file.copy(layerfile, layerpath_new)
+        }
+        else {
+          print(paste("File skipped due to layertype/fileextension mismatch:", basename(layerfile)))
+        }
+        # Remove old layer files if requested and if file was successfully copied
+        if(remove_old && file.exists(layerpath_new)){
             file.remove(layerfile)
-          }
         }
       }
     }
   }
-  # Save xml, zip everything back up
+  # Save xml
   print("Reassembling QGIS project file...")
   write_xml(project_xml, project_new, options = "as_xml")
-  zip(paste(output_dir, "/", project_file, sep=""), files = c(project_new, list.files(path = output_dir, pattern="_styles.db", full.names = TRUE)), extras = "--junk-paths")
+
+  # Get stylefile and zip everything back up
+  style_file <- list.files(path = output_dir, pattern="_styles.db", full.names = TRUE)
+  zip(paste(output_dir, "/", project_file, sep=""), files = c(project_new, style_file), extras = "--junk-paths")
+
   # Delete the friends we made along the way
-  file.remove(c(project_copy, project_new, list.files(path = output_dir, pattern="_styles.db", full.names = TRUE)))
-  # Remove of file if requested and if file was successfully copied
-  if(remove_old){
-    if(file.exists(paste(output_dir, "/", project_file, sep=""))){
+  file.remove(c(project_copy, project_new, style_file))
+
+  # Remove old QGZ file if requested and if file was successfully copied
+  if(remove_old && file.exists(file.path(output_dir, project_file))){
       file.remove(input_path)
-    }
   }
 }
